moved the extraction of currencies with chainlink oracles to advanced-logic as it's only relevant there and added tests to verify the new logic.
